refacto and clean code
