+++ src/components/search/search-bar.tsx
@@ -4,12 +4,10 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-import { RefObject, useCallback, useRef } from 'react';
+import { RefObject, useCallback } from 'react';
 import {
     ElementSearchInput,
     ElementSearchInputProps,
-    ElementType,
-    fetchDirectoryContent,
     Paginated,
     useElementSearch,
     useSnackMessage,
@@ -17,13 +15,13 @@ import {
 import { useDispatch, useSelector } from 'react-redux';
 import { TextFieldProps } from '@mui/material';
 import { searchElementsInfos } from '../../utils/rest-api';
-import { setSearchedElement, setSelectedDirectory, setTreeData } from '../../redux/actions';
-import { updatedTree } from '../treeview-utils';
+import { setSearchedElement } from '../../redux/actions';
 import { SearchItem } from './search-item';
-import { AppState, ElementAttributesES, IDirectory, ITreeData } from '../../redux/types';
+import { AppState, ElementAttributesES } from '../../redux/types';
 import { SearchBarRenderInput } from './search-bar-render-input';
 import { AppDispatch } from '../../redux/store';
 import { SearchBarPaperDisplayedElementWarning } from './search-bar-displayed-element-warning';
+import { useDirectoryPathLoader } from '../../hooks/use-directory-path-loader';
 
 export interface SearchBarProps {
     inputRef: RefObject<TextFieldProps>;
@@ -32,10 +30,7 @@ export interface SearchBarProps {
 export function SearchBar({ inputRef }: Readonly<SearchBarProps>) {
     const dispatch = useDispatch<AppDispatch>();
     const { snackError } = useSnackMessage();
-    const treeData = useSelector((state: AppState) => state.treeData);
-    const treeDataRef = useRef<ITreeData>(null);
     const selectedDirectory = useSelector((state: AppState) => state.selectedDirectory);
-    treeDataRef.current = treeData;
 
     const fetchElementsPageable: (newSearchTerm: string) => Promise<Paginated<ElementAttributesES>> = useCallback(
         (newSearchTerm) => searchElementsInfos(newSearchTerm, selectedDirectory?.elementUuid),
@@ -44,7 +39,7 @@ export function SearchBar({ inputRef }: Readonly<SearchBarProps>) {
     const { elementsFound, isLoading, searchTerm, updateSearchTerm, totalElements } = useElementSearch({
         fetchElements: fetchElementsPageable,
     });
-
+    const { loadPath, handleDispatchDirectory } = useDirectoryPathLoader();
     const renderOptionItem = useCallback<ElementSearchInputProps<ElementAttributesES>['renderElement']>(
         (props) => {
             const { element, inputValue } = props;
@@ -55,67 +50,25 @@ export function SearchBar({ inputRef }: Readonly<SearchBarProps>) {
         [elementsFound]
     );
 
-    const updateMapData = useCallback(
-        (nodeId: string, children: IDirectory[]) => {
-            if (!treeDataRef.current) {
-                return;
-            }
-            const [newRootDirectories, newMapData] = updatedTree(
-                treeDataRef.current.rootDirectories,
-                treeDataRef.current.mapData,
-                nodeId,
-                children
-            );
-            dispatch(
-                setTreeData({
-                    rootDirectories: newRootDirectories,
-                    mapData: newMapData,
-                    initialized: true,
-                })
-            );
-        },
-        [dispatch]
-    );
-
-    const handleDispatchDirectory = useCallback(
-        (elementUuidPath: string | undefined) => {
-            if (treeDataRef.current && elementUuidPath !== undefined) {
-                dispatch(setSelectedDirectory(treeDataRef.current.mapData[elementUuidPath]));
-            }
-        },
-        [dispatch]
-    );
-
     const handleMatchingElement = useCallback<ElementSearchInputProps<ElementAttributesES>['onSelectionChange']>(
         async (data) => {
             const matchingElement = elementsFound.find((element) => element === data);
-            if (matchingElement !== undefined) {
-                const elementUuidPath = matchingElement?.pathUuid;
-                try {
-                    // eslint-disable-next-line no-restricted-syntax -- usage of async/await syntax
-                    for (const uuid of elementUuidPath) {
-                        // eslint-disable-next-line no-await-in-loop -- it's wanted because we don't want to download in parallel
-                        const resources = await fetchDirectoryContent(uuid);
-                        updateMapData(
-                            uuid,
-                            resources.filter((res) => res.type === ElementType.DIRECTORY) as IDirectory[]
-                        );
-                    }
-                } catch (error: any) {
-                    snackError({
-                        messageTxt: error.message,
-                        headerId: 'pathRetrievingError',
-                    });
-                }
-                const lastElement = elementUuidPath.pop();
-
-                dispatch(setSearchedElement(data));
-                if (lastElement !== selectedDirectory?.elementUuid) {
-                    handleDispatchDirectory(lastElement);
-                }
+            if (!matchingElement) return;
+            try {
+                await loadPath(matchingElement.pathUuid);
+            } catch (error: any) {
+                snackError({
+                    messageTxt: error.message,
+                    headerId: 'pathRetrievingError',
+                });
+            }
+            const lastElement = matchingElement.pathUuid.at(-1);
+            dispatch(setSearchedElement(data));
+            if (lastElement !== selectedDirectory?.elementUuid) {
+                handleDispatchDirectory(lastElement);
             }
         },
-        [selectedDirectory?.elementUuid, handleDispatchDirectory, updateMapData, snackError, dispatch, elementsFound]
+        [elementsFound, loadPath, selectedDirectory?.elementUuid, handleDispatchDirectory, snackError, dispatch]
     );
 
     const displayComponent = useCallback<NonNullable<ElementSearchInputProps<ElementAttributesES>['PaperComponent']>>(
